+++ docs/changelog.rst
@@ -12,6 +12,10 @@ Changelog for django-SHOP
 * If an anonymous customer logs in, his current cart is merged with a cart, which has previously
   been created. This has been adopted to re-use the method Product.is_in_cart()
   in and finds it's Merge the contents of the other cart into this one, afterwards delete it.
+* Moved field ``salutation`` from :class:`shop.models.customer.BaseCustomer` into the merchant
+  implementation. If your project does not use the provided default customer model
+  :class:`shop.models.defaults.customer.Customer`, then you should add the ``salutation`` field
+  to your implementation of the Customer model, if that makes sense in your use-case.
 
 
 0.9.3